+++ src/test/java/de/samply/template/ConverterTemplateTest.java
@@ -65,6 +65,9 @@ class ConverterTemplateTest {
     if (random.nextInt(2) == 0){
       attributeTemplate.setOperation(Operation.EXTRACT_RELATIVE_ID.name());
     }
+    if (random.nextInt(2) == 0){
+      attributeTemplate.setConditionFhirPath("condition.path"+i);
+    }
 
     return attributeTemplate;
   }